+++ WhistleCore/Sources/WhistleCore/AppDefaults.swift
@@ -26,7 +26,6 @@ public enum AppDefaults {
         public static let appLockReauthOnForeground = "fmf.appLockReauthOnForeground"
         public static let lastEventTimestamp = "fmf.lastEventTimestamp"
         public static let processedEventIds = "fmf.processedEventIds"
-        public static let pendingLeaveRequests = "fmf.pendingLeaveRequests"
         public static let pendingGiftWrapEventIds = "fmf.pendingGiftWrapEventIds"
         public static let pausedGroupIds = "fmf.pausedGroupIds"
         public static let keyRotationIntervalDays = "fmf.keyRotationIntervalDays"