+++ android/shared/src/test/java/org/findmyfam/shared/AppDefaultsTest.kt
@@ -45,7 +45,6 @@ class AppDefaultsTest {
             AppDefaults.Keys.appLockReauthOnForeground,
             AppDefaults.Keys.lastEventTimestamp,
             AppDefaults.Keys.processedEventIds,
-            AppDefaults.Keys.pendingLeaveRequests,
             AppDefaults.Keys.pendingGiftWrapEventIds,
             AppDefaults.Keys.keyRotationIntervalDays
         )